+++ govtool/frontend/src/components/organisms/RegisterAsDRepSteps/RegisterAsDRepForm.tsx
@@ -70,6 +70,7 @@ export const RegisterAsDRepForm = ({
           placeholder={Placeholders.LINK}
           name={`links.${index}.link`}
           rules={Rules.LINK}
+          data-testid="link-input"
         />
       )),
     [errors, links],
@@ -102,6 +103,7 @@ export const RegisterAsDRepForm = ({
         name="dRepName"
         rules={Rules.DREP_NAME}
         placeholder={t("forms.registerAsDRep.dRepNamePlaceholder")}
+        dataTestId="name-input"
       />
       <Spacer y={isMobile ? 5 : 6} />
       <Box textAlign="center">
@@ -119,6 +121,7 @@ export const RegisterAsDRepForm = ({
         name="email"
         placeholder={t("forms.registerAsDRep.emailPlaceholder")}
         rules={Rules.EMAIL}
+        data-testid="email-input"
       />
       <Spacer y={3} />
       <ControlledField.TextArea
@@ -128,6 +131,7 @@ export const RegisterAsDRepForm = ({
         placeholder={t("forms.registerAsDRep.bioPlaceholder")}
         helpfulText={t("forms.registerAsDRep.bioHelpfulText")}
         rules={Rules.BIO}
+        data-testid="bio-input"
       />
       <Spacer y={4} />
       <p
@@ -149,7 +153,12 @@ export const RegisterAsDRepForm = ({
       <Spacer y={3} />
       {renderLinks()}
       {links?.length < MAX_NUMBER_OF_LINKS ? (
-        <Button onClick={addLink} size="extraLarge" variant="text">
+        <Button
+          onClick={addLink}
+          size="extraLarge"
+          variant="text"
+          data-testid="add-link-button"
+        >
           {t("addLink")}
         </Button>
       ) : null}