+++ govtool/frontend/src/i18n/locales/en.ts
@@ -245,19 +245,31 @@ export const en = {
       },
     },
     dRepDirectory: {
-      abstainCardDescription: "Select this to vote ABSTAIN to every vote.",
-      abstainCardTitle: "Abstain from Every Vote",
+      abstainCardDefaultDescription:
+        "Select this to vote ABSTAIN to every vote.",
+      abstainCardDefaultTitle: "Abstain from Every Vote",
       automatedVotingOptions: "Automated Voting Options",
       editBtn: "Edit DRep data",
+      delegatedToAbstainTitle: "You have delegated ₳{{ada}} to “Abstain”",
+      delegatedToNoConfidenceTitle:
+        "You have delegated ₳{{ada}} to “No Confidence”",
+      delegatedToAbstainDescription:
+        "You have selected to apply your Voting Power to Abstain on every vote.",
+      delegatedToNoConfidenceDescription:
+        "You have selected to apply your Voting Power to No Confidence on every vote.",
       delegationOptions: "Delegation Options",
+      directVoter: "Direct Voter",
       filterTitle: "DRep Status",
+      goToDRepDirectory: "Go to DRep Directory",
       meAsDRep: "This  DRep ID is connected  to your wallet",
       myDelegation: "You have delegated <strong>₳ {{ada}}</strong> to:",
+      myDelegationToYourself:
+        "You have delegated <strong>₳ {{ada}}</strong> to yourself:",
       myDRep: "This is your DRep",
       listTitle: "Find a DRep",
-      noConfidenceDescription:
+      noConfidenceDefaultDescription:
         "Select this to signal no confidence in the current constitutional committee by voting NO on every proposal and voting YES to no confidence proposals",
-      noConfidenceTitle: "Signal No Confidence on Every Vote",
+      noConfidenceDefaultTitle: "Signal No Confidence on Every Vote",
       noResultsForTheSearchTitle: "No DReps found",
       noResultsForTheSearchDescription: "Please try a different search",
       title: "DRep Directory",