+++ govtool/frontend/src/types/automatedVotingOptions.ts
@@ -1,0 +1,9 @@
+export enum AutomatedVotingOptionCurrentDelegation {
+  drep_always_abstain = "drep_always_abstain",
+  drep_always_no_confidence = "drep_always_no_confidence",
+}
+
+export enum AutomatedVotingOptionDelegationId {
+  abstain = "abstain",
+  no_confidence = "no_confidence",
+}