+++ govtool/packages/submission-tool/jest.setup.js
@@ -1,0 +1,5 @@
+require("@testing-library/jest-dom");
+
+const { TextDecoder, TextEncoder } = require("text-encoding");
+global.TextDecoder = TextDecoder;
+global.TextEncoder = TextEncoder;